refactor: remove v1 history and rename v2 history to just history
