+++ prisma/schema.prisma
@@ -25,18 +25,16 @@ model Token {
 }
 
 model Pair {
-  id                          Int                               @id @default(autoincrement())
-  address                     String                            @unique
-  token0                      Token                             @relation("Pair_Token0", fields: [t0], references: [id])
-  t0                          Int
-  token1                      Token                             @relation("Pair_Token1", fields: [t1], references: [id])
-  t1                          Int
-  liquidityInfo               PairLiquidityInfo?                @relation("Pair_Info")
-  synchronized                Boolean
-  liquidityInfoHistory        PairLiquidityInfoHistory[]
-  liquidityInfoHistoryV2      PairLiquidityInfoHistoryV2[]
-  liquidityInfoHistoryError   PairLiquidityInfoHistoryError[]
-  liquidityInfoHistoryV2Error PairLiquidityInfoHistoryV2Error[]
+  id                        Int                             @id @default(autoincrement())
+  address                   String                          @unique
+  token0                    Token                           @relation("Pair_Token0", fields: [t0], references: [id])
+  t0                        Int
+  token1                    Token                           @relation("Pair_Token1", fields: [t1], references: [id])
+  t1                        Int
+  liquidityInfo             PairLiquidityInfo?              @relation("Pair_Info")
+  synchronized              Boolean
+  liquidityInfoHistory      PairLiquidityInfoHistory[]
+  liquidityInfoHistoryError PairLiquidityInfoHistoryError[]
 }
 
 model PairLiquidityInfo {
@@ -49,34 +47,6 @@ model PairLiquidityInfo {
 }
 
 model PairLiquidityInfoHistory {
-  id             Int      @id @default(autoincrement())
-  pair           Pair     @relation(fields: [pairId], references: [id])
-  pairId         Int
-  totalSupply    String
-  reserve0       String
-  reserve1       String
-  height         Int
-  microBlockHash String
-  microBlockTime BigInt
-  updatedAt      DateTime @default(now()) @updatedAt
-
-  @@unique(name: "pairIdMicroBlockHashUniqueIndex", [pairId, microBlockHash])
-}
-
-model PairLiquidityInfoHistoryError {
-  id             Int      @id @default(autoincrement())
-  pair           Pair     @relation(fields: [pairId], references: [id])
-  pairId         Int
-  microBlockHash String
-  error          String
-  timesOccurred  Int      @default(1)
-  createdAt      DateTime @default(now())
-  updatedAt      DateTime @default(now()) @updatedAt
-
-  @@unique(name: "pairIdMicroBlockHashErrorUniqueIndex", [pairId, microBlockHash, error])
-}
-
-model PairLiquidityInfoHistoryV2 {
   id               Int      @id @default(autoincrement())
   pair             Pair     @relation(fields: [pairId], references: [id])
   pairId           Int
@@ -98,7 +68,7 @@ model PairLiquidityInfoHistoryV2 {
   @@unique(name: "pairIdMicroBlockHashTxHashLogIndexUniqueIndex", [pairId, microBlockHash, transactionHash, logIndex])
 }
 
-model PairLiquidityInfoHistoryV2Error {
+model PairLiquidityInfoHistoryError {
   id              Int      @id @default(autoincrement())
   pair            Pair     @relation(fields: [pairId], references: [id])
   pairId          Int
