+++ src/tasks/pair-liquidity-info-history-validator/pair-liquidity-info-history-validator.service.ts
@@ -1,20 +1,27 @@
 import { Injectable, Logger } from '@nestjs/common';
-import { uniq } from 'lodash';
+import { groupBy, last, map } from 'lodash';
 
 import { MdwHttpClientService } from '@/clients/mdw-http-client.service';
-import { MicroBlockHash } from '@/clients/sdk-client.model';
+import {
+  ContractAddress,
+  contractAddrToAccountAddr,
+  MicroBlockHash,
+} from '@/clients/sdk-client.model';
 import { SdkClientService } from '@/clients/sdk-client.service';
 import { PairLiquidityInfoHistoryDbService } from '@/database/pair-liquidity-info-history/pair-liquidity-info-history-db.service';
+import { decimalToBigInt } from '@/lib/utils';
 
 @Injectable()
 export class PairLiquidityInfoHistoryValidatorService {
   constructor(
-    private mdwClient: MdwHttpClientService,
     private pairLiquidityInfoHistoryDb: PairLiquidityInfoHistoryDbService,
+    private mdwClient: MdwHttpClientService,
     private sdkClient: SdkClientService,
   ) {}
 
   readonly logger = new Logger(PairLiquidityInfoHistoryValidatorService.name);
+
+  readonly VALIDATION_WINDOW_BLOCKS = 20;
 
   async validate() {
     this.logger.log(`Started validating pair liquidity info history.`);
@@ -22,33 +29,60 @@ export class PairLiquidityInfoHistoryValidatorService {
     // Get current height
     const currentHeight = await this.sdkClient.getHeight();
 
-    // Get all liquidity entries greater or equal the current height minus 20
-    const liquidityEntriesWithinHeightSorted =
-      await this.pairLiquidityInfoHistoryDb.getWithinHeightSorted(
-        currentHeight - 20,
-      );
-
-    // Get unique keyBlocks from entries
-    const uniqueHeights = uniq(
-      liquidityEntriesWithinHeightSorted.map((e) => e.height),
+    // Get all liquidity entries greater or equal the current height minus VALIDATION_WINDOW_BLOCKS
+    // and take the last entry of every microBlock to get the final reserve in that microBlock
+    const liquidityEntriesWithinHeightSorted = map(
+      groupBy(
+        await this.pairLiquidityInfoHistoryDb.getWithinHeightSortedWithPair(
+          currentHeight - this.VALIDATION_WINDOW_BLOCKS,
+        ),
+        'microBlockHash',
+      ),
+      (group) => last(group)!,
     );
 
-    // Fetch microBlocks for these unique keyBlocks from mdw
-    const microBlockHashsOnMdw = (
-      await Promise.all(
-        uniqueHeights.map((h) => this.mdwClient.getKeyBlockMicroBlocks(h)),
-      )
-    )
-      .flat()
-      .map((mb) => mb.hash);
-
-    // If a local microBlock is not contained in the mdw, delete this block and all newer entries
+    // If the reserves of a local microBlock do not match with the data from the middleware or the block does not exist,
+    // delete all logs in this block and all newer entries
     let numDeleted = 0;
     for (const liquidityEntry of liquidityEntriesWithinHeightSorted) {
+      let isError = false;
+      let mdwReserve0: bigint | undefined;
+      let mdwReserve1: bigint | undefined;
+
+      try {
+        // reserve0 is the balance of the pair contract's account of token0
+        mdwReserve0 = BigInt(
+          (
+            await this.mdwClient.getAccountBalanceForContractAtMicroBlockHash(
+              liquidityEntry.pair.token0.address as ContractAddress,
+              contractAddrToAccountAddr(
+                liquidityEntry.pair.address as ContractAddress,
+              ),
+              liquidityEntry.microBlockHash as MicroBlockHash,
+            )
+          ).amount,
+        );
+
+        // reserve1 is the balance of the pair contract's account of token1
+        mdwReserve1 = BigInt(
+          (
+            await this.mdwClient.getAccountBalanceForContractAtMicroBlockHash(
+              liquidityEntry.pair.token1.address as ContractAddress,
+              contractAddrToAccountAddr(
+                liquidityEntry.pair.address as ContractAddress,
+              ),
+              liquidityEntry.microBlockHash as MicroBlockHash,
+            )
+          ).amount,
+        );
+      } catch (e) {
+        this.logger.error(e);
+        isError = true;
+      }
       if (
-        !microBlockHashsOnMdw.includes(
-          liquidityEntry.microBlockHash as MicroBlockHash,
-        )
+        isError ||
+        decimalToBigInt(liquidityEntry.reserve0) !== mdwReserve0 ||
+        decimalToBigInt(liquidityEntry.reserve1) !== mdwReserve1
       ) {
         numDeleted = (
           await this.pairLiquidityInfoHistoryDb.deleteFromMicroBlockTime(
