+++ src/tasks/tasks.service.spec.ts
@@ -6,50 +6,42 @@ import { MdwHttpClientService } from '@/clients/mdw-http-client.service';
 import { SdkClientService } from '@/clients/sdk-client.service';
 import { PairDbService } from '@/database/pair/pair-db.service';
 import { PairLiquidityInfoHistoryDbService } from '@/database/pair-liquidity-info-history/pair-liquidity-info-history-db.service';
-import { PairLiquidityInfoHistoryV2DbService } from '@/database/pair-liquidity-info-history/pair-liquidity-info-history-v2-db.service';
 import { PairLiquidityInfoHistoryErrorDbService } from '@/database/pair-liquidity-info-history-error/pair-liquidity-info-history-error-db.service';
-import { PairLiquidityInfoHistoryV2ErrorDbService } from '@/database/pair-liquidity-info-history-error/pair-liquidity-info-history-v2-error-db.service';
 import { PrismaService } from '@/database/prisma.service';
 import { PairLiquidityInfoHistoryImporterService } from '@/tasks/pair-liquidity-info-history-importer/pair-liquidity-info-history-importer.service';
-import { PairLiquidityInfoHistoryImporterV2Service } from '@/tasks/pair-liquidity-info-history-importer/pair-liquidity-info-history-importer-v2.service';
 import { PairLiquidityInfoHistoryValidatorService } from '@/tasks/pair-liquidity-info-history-validator/pair-liquidity-info-history-validator.service';
-import { PairLiquidityInfoHistoryValidatorV2Service } from '@/tasks/pair-liquidity-info-history-validator/pair-liquidity-info-history-validator-v2.service';
 import { TasksService } from '@/tasks/tasks.service';
 
 describe('TasksService', () => {
   let tasksService: TasksService;
-  let pairLiquidityInfoHistoryImporterService: PairLiquidityInfoHistoryImporterV2Service;
-  let pairLiquidityInfoHistoryValidatorService: PairLiquidityInfoHistoryValidatorV2Service;
+  let pairLiquidityInfoHistoryImporterService: PairLiquidityInfoHistoryImporterService;
+  let pairLiquidityInfoHistoryValidatorService: PairLiquidityInfoHistoryValidatorService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       providers: [
         TasksService,
         PairLiquidityInfoHistoryImporterService,
-        PairLiquidityInfoHistoryImporterV2Service,
         PairLiquidityInfoHistoryValidatorService,
-        PairLiquidityInfoHistoryValidatorV2Service,
         HttpService,
         CoinmarketcapClientService,
         MdwHttpClientService,
         SdkClientService,
         PairDbService,
         PairLiquidityInfoHistoryDbService,
-        PairLiquidityInfoHistoryV2DbService,
         PairLiquidityInfoHistoryErrorDbService,
-        PairLiquidityInfoHistoryV2ErrorDbService,
         PrismaService,
       ],
     }).compile();
 
     tasksService = module.get<TasksService>(TasksService);
     pairLiquidityInfoHistoryImporterService =
-      module.get<PairLiquidityInfoHistoryImporterV2Service>(
-        PairLiquidityInfoHistoryImporterV2Service,
+      module.get<PairLiquidityInfoHistoryImporterService>(
+        PairLiquidityInfoHistoryImporterService,
       );
     pairLiquidityInfoHistoryValidatorService =
-      module.get<PairLiquidityInfoHistoryValidatorV2Service>(
-        PairLiquidityInfoHistoryValidatorV2Service,
+      module.get<PairLiquidityInfoHistoryValidatorService>(
+        PairLiquidityInfoHistoryValidatorService,
       );
   });
 
@@ -59,7 +51,7 @@ describe('TasksService', () => {
         .spyOn(pairLiquidityInfoHistoryImporterService, 'import')
         .mockResolvedValue();
 
-      await tasksService.runPairLiquidityInfoHistoryImporterV2();
+      await tasksService.runPairLiquidityInfoHistoryImporter();
       expect(pairLiquidityInfoHistoryImporterService.import).toHaveBeenCalled();
       expect(tasksService.isRunning).toBe(false);
     });
@@ -67,7 +59,7 @@ describe('TasksService', () => {
     it('should not run if a task is running already', async () => {
       tasksService.setIsRunning(true);
       jest.spyOn(pairLiquidityInfoHistoryImporterService, 'import');
-      await tasksService.runPairLiquidityInfoHistoryImporterV2();
+      await tasksService.runPairLiquidityInfoHistoryImporter();
       expect(
         pairLiquidityInfoHistoryImporterService.import,
       ).not.toHaveBeenCalled();
@@ -80,7 +72,7 @@ describe('TasksService', () => {
         .mockRejectedValue(error);
       jest.spyOn(pairLiquidityInfoHistoryImporterService.logger, 'error');
 
-      await tasksService.runPairLiquidityInfoHistoryImporterV2();
+      await tasksService.runPairLiquidityInfoHistoryImporter();
       expect(
         pairLiquidityInfoHistoryImporterService.logger.error,
       ).toHaveBeenCalledWith(`Import failed. ${error}`);
@@ -94,7 +86,7 @@ describe('TasksService', () => {
         .spyOn(pairLiquidityInfoHistoryValidatorService, 'validate')
         .mockResolvedValue();
 
-      await tasksService.runPairLiquidityInfoHistoryValidatorV2();
+      await tasksService.runPairLiquidityInfoHistoryValidator();
       expect(
         pairLiquidityInfoHistoryValidatorService.validate,
       ).toHaveBeenCalled();
@@ -105,7 +97,7 @@ describe('TasksService', () => {
 
       jest.spyOn(pairLiquidityInfoHistoryValidatorService, 'validate');
 
-      await tasksService.runPairLiquidityInfoHistoryValidatorV2();
+      await tasksService.runPairLiquidityInfoHistoryValidator();
       expect(
         pairLiquidityInfoHistoryValidatorService.validate,
       ).not.toHaveBeenCalled();
@@ -118,7 +110,7 @@ describe('TasksService', () => {
         .mockRejectedValue(error);
       jest.spyOn(pairLiquidityInfoHistoryValidatorService.logger, 'error');
 
-      await tasksService.runPairLiquidityInfoHistoryValidatorV2();
+      await tasksService.runPairLiquidityInfoHistoryValidator();
       expect(
         pairLiquidityInfoHistoryValidatorService.logger.error,
       ).toHaveBeenCalledWith(`Validation failed. ${error}`);
